+++ lua/dap/constants.lua
@@ -1,4 +1,5 @@
 
 return {
-  ns_pos = 'dap_pos'
+  ns_pos = 'dap_pos',
+  ns_breakpoints = 'dap_breakpoints',
 }
