+++ lua/dap/session.lua
@@ -5,13 +5,17 @@ local log = require('dap.log').create_logger('dap.log')
 local reloadable = require('dap.reloadable')
 local repl = require('dap.repl')
 local rpc = require('dap.rpc')
+local signs = require('dap.signs')
 local ui = require('dap.ui')
 local utils = require('dap.utils')
 
 local ns_pos = require('dap.constants').ns_pos
 
-local non_empty = utils.non_empty
 local convert_nil = utils.convert_nil
+local index_of = utils.index_of
+local non_empty = utils.non_empty
+
+local bp_info = reloadable.table('BpInfo')
 
 local function launch_external_terminal(terminal, args)
   local handle
@@ -57,6 +61,87 @@ local function session_defaults(opts)
   }
 end
 
+local function with_win(win, fn, ...)
+  local cur_win = api.nvim_get_current_win()
+  api.nvim_set_current_win(win)
+  local ok, err = pcall(fn, ...)
+  api.nvim_set_current_win(cur_win)
+  assert(ok, err)
+end
+
+local function jump_to_location(bufnr, line, column)
+  local ok, failure = pcall(vim.fn.sign_place, 0, ns_pos, 'DapStopped', bufnr, { lnum = line; priority = 12 })
+  if not ok then
+    print(failure)
+  end
+  -- vscode-go sends columns with 0
+  -- That would cause a "Column value outside range" error calling nvim_win_set_cursor
+  -- nvim-dap says "columnsStartAt1 = true" on initialize :/
+  if column == 0 then
+    column = 1
+  end
+  for _, win in pairs(api.nvim_tabpage_list_wins(0)) do
+    if api.nvim_win_get_buf(win) == bufnr then
+      api.nvim_win_set_cursor(win, { line, column - 1 })
+      with_win(win, api.nvim_command, 'normal zv')
+      return
+    end
+  end
+  -- Buffer isn't active in any window; use the first window that is not special
+  -- (Don't want to move to code in the REPL...)
+  for _, win in pairs(api.nvim_tabpage_list_wins(0)) do
+    local winbuf = api.nvim_win_get_buf(win)
+    if api.nvim_buf_get_option(winbuf, 'buftype') == '' then
+      local bufchanged, _ = pcall(api.nvim_win_set_buf, win, bufnr)
+      if bufchanged then
+        api.nvim_win_set_cursor(win, { line, column - 1 })
+        with_win(win, api.nvim_command, 'normal zv')
+        return
+      end
+    end
+  end
+end
+
+
+local function jump_to_frame(cur_session, frame, preserve_focus_hint)
+  local source = frame.source
+  if not source then
+    print('Source not available, cannot jump to frame')
+    return
+  end
+  vim.fn.sign_unplace(ns_pos)
+  if preserve_focus_hint or frame.line < 0 then
+    return
+  end
+  if not source.sourceReference or source.sourceReference == 0 then
+    if not source.path then
+      print('Source path not available, cannot jump to frame')
+      return
+    end
+    local scheme = source.path:match('^([a-z]+)://.*')
+    local bufnr
+    if scheme then
+      bufnr = vim.uri_to_bufnr(source.path)
+    else
+      bufnr = vim.uri_to_bufnr(vim.uri_from_fname(source.path))
+    end
+    vim.fn.bufload(bufnr)
+    jump_to_location(bufnr, frame.line, frame.column)
+  else
+    local params = {
+      source = source,
+      sourceReference = source.sourceReference
+    }
+    cur_session:request('source', params, function(err, response)
+      assert(not err, vim.inspect(err))
+
+      local buf = api.nvim_create_buf(false, true)
+      api.nvim_buf_set_lines(buf, 0, -1, false, vim.split(response.content, '\n'))
+      jump_to_location(buf, frame.line, frame.column)
+    end)
+  end
+end
+
 
 local Session = {}
 Session.__index = Session
@@ -94,12 +179,13 @@ function Session:spawn(adapter, opts)
   }
 
   stdout:read_start(rpc.create_read_loop(function(body)
-    self:handle_body(body)
+    o:handle_body(body)
   end))
 
   stderr:read_start(function(err, chunk)
     assert(not err, err)
-    local _ = chunk and log.error() and log.error("stderr", adapter, chunk)
+    if not chunk then return end
+    local _ = log.error() and log.error("stderr", adapter, chunk)
   end)
 
   return o
@@ -119,7 +205,7 @@ function Session:connect(host, port, opts)
   client:connect(host or '127.0.0.1', tonumber(port), function(err)
     if (err) then print(err) end
   end)
-  client:read_start(rpc.create_read_loop(function(body) self:handle_body(body) end))
+  client:read_start(rpc.create_read_loop(function(body) o:handle_body(body) end))
   return o
 end
 
@@ -403,6 +489,321 @@ function Session:_step(step)
   end)
 end
 
+function Session:event_stopped(stopped)
+  if self.stopped_thread_id then
+    log.debug('Thread stopped, but another thread is already stopped, telling thread to continue')
+    self:request('continue', { threadId = stopped.threadId })
+    return
+  end
+  self.stopped_thread_id = stopped.threadId
+  self:request('threads', nil, function(err0, threads_resp)
+    if err0 then
+      print('Error retrieving threads: ' .. err0.message)
+      return
+    end
+    local threads = {}
+    self.threads = threads
+    for _, thread in pairs(threads_resp.threads) do
+      threads[thread.id] = thread
+    end
+
+    if stopped.reason == 'exception' then
+      self:_show_exception_info()
+    end
+
+    self:request('stackTrace', { threadId = stopped.threadId; }, function(err1, frames_resp)
+      if err1 then
+        print('Error retrieving stack traces: ' .. err1.message)
+        return
+      end
+      local frames = {}
+      local current_frame = nil
+      self.current_frame = nil
+      threads[stopped.threadId].frames = frames
+      for _, frame in pairs(frames_resp.stackFrames) do
+        if not current_frame and frame.source and frame.source.path then
+          current_frame = frame
+          self.current_frame = frame
+        end
+        table.insert(frames, frame)
+      end
+      if not current_frame then
+        if #frames > 0 then
+          current_frame = frames[1]
+          self.current_frame = current_frame
+        else
+          return
+        end
+      end
+      local preserve_focus
+      if stopped.reason ~= 'pause' then
+        preserve_focus = stopped.preserveFocusHint
+      end
+      jump_to_frame(self, current_frame, preserve_focus)
+      self:_request_scopes(current_frame)
+    end)
+  end)
+end
+
+function Session:event_terminated()
+  self:close()
+  if self.handlers.after then
+    self.handlers.after()
+  end
+
+  -- TODO: Ugly hax
+  --    Could probably do this with custom event handlers?
+  require('dap')._set_current_session(nil)
+end
+
+
+function Session.event_exited()
+end
+
+function Session.event_module()
+end
+
+function Session.event_process()
+end
+
+function Session.event_thread()
+end
+
+function Session.event_continued()
+end
+
+function Session.event_output(_, body)
+  repl.append(body.output, '$')
+end
+
+function Session:_request_scopes(current_frame)
+  self:request('scopes', { frameId = current_frame.id }, function(_, scopes_resp)
+    if not scopes_resp or not scopes_resp.scopes then return end
+
+    current_frame.scopes = {}
+    for _, scope in pairs(scopes_resp.scopes) do
+
+      table.insert(current_frame.scopes, scope)
+      if not scope.expensive then
+        self:request('variables', { variablesReference = scope.variablesReference }, function(_, variables_resp)
+          if not variables_resp then return end
+
+          scope.variables = utils.to_dict(
+            variables_resp.variables,
+            function(v) return v.name end
+          )
+        end)
+      end
+    end
+  end)
+end
+
+--- Goto specified line (source and col are optional)
+function Session:_goto(line, source, col)
+  local frame = self.current_frame
+  if not frame then
+    return
+  end
+  if not self.capabilities.supportsGotoTargetsRequest then
+    print("Debug Adapter doesn't support GotoTargetRequest")
+    return
+  end
+  self:request('gotoTargets',  {source = source or frame.source, line = line, col = col}, function(err, response)
+    if err then
+      print('Error getting gotoTargets: ' .. err.message)
+      return
+    end
+    if not response or not response.targets then
+      print("No goto targets available. Can't execute goto")
+      return
+    end
+    local params = {threadId = self.stopped_thread_id, targetId = response.targets[1].id }
+    self:request('goto', params, function(err1, _)
+      if err1 then
+        print('Error executing goto: ' .. err1.message)
+      end
+    end)
+  end)
+
+end
+
+
+function Session:_frame_set(frame)
+  if not frame then
+    return
+  end
+  self.current_frame = frame
+  jump_to_frame(self, frame, false)
+  self:_request_scopes(frame)
+end
+
+
+function Session:_frame_delta(delta)
+  if not self.stopped_thread_id then
+    print('Cannot move frame if not stopped')
+    return
+  end
+  local frames = self.threads[self.stopped_thread_id].frames
+  assert(frames, 'Stopped thread must have frames')
+  local current_frame_index = index_of(frames, function(i) return i.id == self.current_frame.id end)
+  assert(current_frame_index, 'id of current frame must be present in frames')
+
+  current_frame_index = current_frame_index + delta
+  if current_frame_index < 1 then
+    current_frame_index = #frames
+  elseif current_frame_index > #frames then
+    current_frame_index = 1
+  end
+  self:_frame_set(frames[current_frame_index])
+end
+
+function Session:_show_exception_info()
+  if not self.capabilities.supportsExceptionInfoRequest then return end
+
+  -- exceptionInfo (https://microsoft.github.io/debug-adapter-protocol/specification#Requests_ExceptionInfo)
+  --- threadId: number
+  self:request('exceptionInfo', {threadId = self.stopped_thread_id}, function(err, response)
+    if err then
+      print("Error getting exception info: "..err.message)
+    end
+
+    -- ExceptionInfoResponse
+    --- exceptionId: string
+    --- description?: string
+    --- breakMode: ExceptionBreakMode (https://microsoft.github.io/debug-adapter-protocol/specification#Types_ExceptionBreakMode)
+    --- details?: ExceptionDetails (https://microsoft.github.io/debug-adapter-protocol/specification#Types_ExceptionDetails)
+    if response then
+      local exception_type = response.details and response.details.typeName
+      local of_type = exception_type and ' of type '..exception_type or ''
+      repl.append('Thread stopped due to exception'..of_type..' ('..response.breakMode..')')
+      if response.description then
+        repl.append('Description: '..response.description)
+      end
+      -- ExceptionDetails (https://microsoft.github.io/debug-adapter-protocol/specification#Types_ExceptionDetails)
+      --- message?: string
+      --- typeName?: string
+      --- fullTypeName?: string
+      --- evaluateName?: string
+      --- stackTrace?: string
+      --- innerException?: ExceptionDetails[]
+      if response.details then
+        if response.details.stackTrace then
+          repl.append("Stack trace:")
+          repl.append(response.details.stackTrace)
+        end
+        if response.details.innerException then
+          repl.append("Inner Exceptions:")
+          for _, e in pairs(response.details.innerException) do
+            repl.append(vim.inspect(e))
+          end
+        end
+      end
+    end
+  end)
+end
+
+function Session:set_breakpoints(bufexpr, on_done)
+  local bp_signs = signs.get_breakpoints(bufexpr)
+
+  local num_bufs = #bp_signs
+  if num_bufs == 0 then
+    on_done()
+    return
+  end
+  for _, buf_bp_signs in pairs(bp_signs) do
+    local breakpoints = {}
+    local bufnr = buf_bp_signs.bufnr
+    for _, bp in pairs(buf_bp_signs.signs) do
+      local bp_entry = bp_info[bp.id] or {}
+      table.insert(breakpoints, {
+        line = bp.lnum;
+        condition = bp_entry.condition;
+        hitCondition = bp_entry.hitCondition;
+        logMessage = bp_entry.logMessage;
+      })
+    end
+    if non_empty(bp_info.condition) and not self.capabilities.supportsConditionalBreakpoints then
+      print("Debug adapter doesn't support breakpoints with conditions")
+    end
+    if non_empty(bp_info.hitCondition) and not self.capabilities.supportsHitConditionalBreakpoints then
+      print("Debug adapter doesn't support breakpoints with hit conditions")
+    end
+    if non_empty(bp_info.logMessage) and not self.capabilities.supportsLogPoints then
+      print("Debug adapter doesn't support log points")
+    end
+    local path = api.nvim_buf_get_name(bufnr)
+    local payload = {
+      source = {
+        path = path;
+        name = vim.fn.fnamemodify(path, ':t')
+      };
+      breakpoints = breakpoints;
+      lines = vim.tbl_map(function(x) return x.line end, breakpoints);
+    }
+    self:request('setBreakpoints', payload, function(err1, resp)
+        if err1 then
+          print("Error setting breakpoints: " .. err1.message)
+        else
+          for _, bp in pairs(resp.breakpoints) do
+            if not bp.verified then
+              log.info('Server rejected breakpoint', bp)
+              signs.remove_breakpoints(bufnr, bp.line)
+            end
+          end
+        end
+        num_bufs = num_bufs - 1
+        if num_bufs == 0 and on_done then
+          on_done()
+        end
+      end
+    )
+  end
+end
+
+function Session:set_exception_breakpoints(filters, exceptionOptions, on_done)
+  if not self.capabilities.exceptionBreakpointFilters then
+      print("Debug adapter doesn't support exception breakpoints")
+      return
+  end
+
+  if filters == 'default' then
+    local default_filters = {}
+    for _, f in pairs(self.capabilities.exceptionBreakpointFilters) do
+      if f.default then
+        table.insert(default_filters, f.filter)
+      end
+    end
+    filters = default_filters
+  end
+
+  if not filters then
+    local possible_filters = {}
+    for _, f in ipairs(self.capabilities.exceptionBreakpointFilters) do
+      table.insert(possible_filters, f.filter)
+    end
+    filters = vim.split(vim.fn.input("Exception breakpoint filters: ", table.concat(possible_filters, ' ')), ' ')
+  end
+
+  if exceptionOptions and not self.capabilities.supportsExceptionOptions then
+    print("Debug adapter does not support ExceptionOptions")
+    return
+  end
+
+  -- setExceptionBreakpoints (https://microsoft.github.io/debug-adapter-protocol/specification#Requests_SetExceptionBreakpoints)
+  --- filters: string[]
+  --- exceptionOptions: exceptionOptions?: ExceptionOptions[] (https://microsoft.github.io/debug-adapter-protocol/specification#Types_ExceptionOptions)
+  self:request(
+    'setExceptionBreakpoints',
+    { filters = filters, exceptionOptions = exceptionOptions },
+    function(err, _)
+      if err then
+        print("Error setting exception breakpoints: "..err.message)
+      end
+      if on_done then
+        on_done()
+      end
+  end)
+end
 
 
 return Session
